Rebrand the filter controls
